✨ feat : 탈퇴사유 도넛차트 완성
